separate campaign_run repository and improve variable naming per PR review
