+++ src/rsvp_service/get_rsvp_status/campaign_repository.py
@@ -1,10 +1,8 @@
 import logging
-
 import boto3
-from boto3.dynamodb.conditions import Attr, Key
+from boto3.dynamodb.conditions import Key
 
 logger = logging.getLogger()
-
 
 class CampaignRepository:
     def __init__(self, table_name):
@@ -12,32 +10,14 @@ class CampaignRepository:
         self.table = self.dynamodb.Table(table_name)
 
     def get_campaign_by_id(self, campaign_id, cohort="8"):
+
         try:
             response = self.table.query(
-                KeyConditionExpression=Key("cohort").eq(cohort)
-                & Key("campaign_id_created_at").begins_with(campaign_id)
+                KeyConditionExpression=Key("cohort").eq(cohort) & 
+                                       Key("campaign_id_created_at").begins_with(campaign_id)
             )
             items = response.get("Items", [])
             return items[0] if items else None
         except Exception as e:
-            logger.error("Error querying campaign: %s", e)
-            raise e
-
-    def get_run(self, campaign_id, run_id):
-        try:
-            response = self.table.get_item(
-                Key={"campaign_id": campaign_id, "run_id": run_id}
-            )
-            return response.get("Item")
-        except Exception as e:
-            logger.error("Error fetching run: %s", e)
-            raise e
-
-    def scan_run_by_id(self, run_id):
-        try:
-            response = self.table.scan(FilterExpression=Attr("run_id").eq(run_id))
-            items = response.get("Items", [])
-            return items[0] if items else None
-        except Exception as e:
-            logger.error("Error scanning run: %s", e)
+            logger.error("Error querying campaign in campaign_repository: %s", e)
             raise e
